🚚 [chatExtras] renamed Module + adjusted settings
