+++ src/modules/chatExtras/i18n/cs_CZ.root.json
@@ -1,8 +1,12 @@
 {
-	"description": "Namísto původního času zobrazí v chatu spolu s časem i datum.",
-	"name": "Datum v chatu",
+	"description": "",
+	"name": "",
 	"settings": {
-		"format": {
+		"chatTime": {
+			"description": "",
+			"title": ""
+		},
+		"chatTimeFormat": {
 			"description": "Zde můžete zvolit formát datumu. Nápovědu naleznete <a class=\"lightbox-open\" href=\"{wiki}/modules/clock.html#configuration\">zde</a>.",
 			"title": "Formát datumu a času"
 		}
