+++ src/modules/chatExtras/i18n/de_DE.root.json
@@ -1,8 +1,12 @@
 {
-	"description": "Tauscht die Uhrzeit im Chat gegen Datum und Uhrzeit aus.",
-	"name": "Chat Zeitstempel",
+	"description": "Fügt ein paar nützliche Verbesserungen am Chat durch",
+	"name": "Chat-Erweiterungen",
 	"settings": {
-		"format": {
+		"chatTime": {
+			"description": "Verändert den Zeitstempel. Das Format kannst du selbstständig anpassen.",
+			"title": "Zeitstempel im Chat anpassen"
+		},
+		"chatTimeFormat": {
 			"description": "Hier kannst du das Format deiner Wahl festlegen. Eine Hilfe hierzu findest du <a class=\"lightbox-open\" href=\"{wiki}/modules/clock.html#konfiguration\">hier</a>.",
 			"title": "Format des Zeitstempel des Chats"
 		}
