+++ src/modules/chatExtras/i18n/fr_FR.root.json
@@ -1,8 +1,12 @@
 {
-	"description": "Affiche la date et l'heure au lieu de l'heure seulement",
-	"name": "Format de date",
+	"description": "",
+	"name": "",
 	"settings": {
-		"format": {
+		"chatTime": {
+			"description": "",
+			"title": ""
+		},
+		"chatTimeFormat": {
 			"description": "Ici vous pouvez choisir le format de votre choix. Pour l'aide c'est <a class=\"lightbox-open\" href=\"{wiki}/modules/clock.html#configuration\">ici</a>.",
 			"title": "Format date et heure"
 		}
