+++ src/modules/chatExtras/i18n/it_IT.root.json
@@ -1,8 +1,12 @@
 {
-	"description": "Mostra la data e l'ora anziché solo l'ora",
-	"name": "Formato della data della chat",
+	"description": "",
+	"name": "",
 	"settings": {
-		"format": {
+		"chatTime": {
+			"description": "",
+			"title": ""
+		},
+		"chatTimeFormat": {
 			"description": "Qui puoi scegliere il formato che preferisci. Puoi trovare aiuto <a class=\"lightbox-open\" href=\"{wiki}/modules/clock.html#configuration\">here</a>.",
 			"title": "Formato della data e dell'ora."
 		}
