+++ src/modules/chatExtras/i18n/nb_NO.root.json
@@ -1,8 +1,12 @@
 {
-	"description": "Viser dato og tid istedenfor kun tid i chatten.",
-	"name": "Utvidet Chat",
+	"description": "",
+	"name": "",
 	"settings": {
-		"format": {
+		"chatTime": {
+			"description": "",
+			"title": ""
+		},
+		"chatTimeFormat": {
 			"description": "Her kan du velge formatet du selv ønsker, du kan finne hjelp <a class=\"lightbox-open\" href=\"{wiki}/modules/clock.html#configuration\">her</a>.",
 			"title": "Formatet til dato og tid"
 		}
