+++ src/modules/chatExtras/i18n/nl_NL.root.json
@@ -1,8 +1,12 @@
 {
-	"description": "Geeft datum en tijd weer in de chat i.p.v. alleen tijd.",
-	"name": "Chat datumweergave",
+	"description": "",
+	"name": "",
 	"settings": {
-		"format": {
+		"chatTime": {
+			"description": "",
+			"title": ""
+		},
+		"chatTimeFormat": {
 			"description": "Hier kunt u het formaat kiezen van uw keuze. Hulp kunt u <a class=\"lightbox-open\" href=\"{wiki}/modules/clock.html#configuratie\">hier</a> vinden.",
 			"title": "Formaat van de datum en tijd"
 		}
