+++ src/modules/chatExtras/settings.ts
@@ -1,9 +1,14 @@
 import { ModuleSettingFunction } from 'typings/Module';
-import { Text } from 'typings/Setting';
+import { Text, Toggle } from 'typings/Setting';
 
 export default (() => ({
-    format: <Text>{
+    chatTime: <Toggle>{
+        type: 'toggle',
+        default: true,
+    },
+    chatTimeFormat: <Text>{
         type: 'text',
         default: 'DD.MM LTS',
+        dependsOn: '.betterChatTime',
     },
 })) as ModuleSettingFunction;
