+++ src/modules/missionHelper/i18n/en_US.json
@@ -2,7 +2,6 @@
 	"diyMission": "This mission seems to be a large scale alliance mission!",
 	"expansions": "Spread | Spreads",
 	"followups": "Follow-up Mission | Follow-up Missions",
-	"subsequent": "Subsequent Mission | Subsequent Missions",
 	"generated_by": "Generated by",
 	"noVehicleRequirements": [
 		"water_needed"
@@ -39,6 +38,7 @@
 	"requirements": {
 		"water_needed": "Water demand (gal)"
 	},
+	"subsequent": "Subsequent Mission | Subsequent Missions",
 	"vehicles": {
 		"captions": {
 			"ambulances": "Ambulance | Ambulances",
