+++ src/modules/missionHelper/settings.ts
@@ -38,12 +38,12 @@ export default ((MODULE_ID: string) => ({
     },
     ...(['de_DE', 'en_US', 'nl_NL'].includes(BUILD_LANG)
         ? {
-            'vehicles.patient_additionals': <Toggle>{
-                type: 'toggle',
-                default: true,
-                dependsOn: '.vehicles.content',
-            },
-        }
+              'vehicles.patient_additionals': <Toggle>{
+                  type: 'toggle',
+                  default: true,
+                  dependsOn: '.vehicles.content',
+              },
+          }
         : null),
     'chances.normal': <Toggle>{
         type: 'toggle',
@@ -56,12 +56,12 @@ export default ((MODULE_ID: string) => ({
     },
     ...(BUILD_LANG !== 'nl_NL'
         ? {
-            'multifunctionals.heavy_rescue_vehicles': <Toggle>{
-                type: 'toggle',
-                default: false,
-                dependsOn: '.vehicles.content',
-            },
-        }
+              'multifunctionals.heavy_rescue_vehicles': <Toggle>{
+                  type: 'toggle',
+                  default: false,
+                  dependsOn: '.vehicles.content',
+              },
+          }
         : null),
     'multifunctionals.battalion_chief_vehicles': <Toggle>{
         type: 'toggle',
@@ -84,21 +84,21 @@ export default ((MODULE_ID: string) => ({
     },
     ...(BUILD_LANG === 'nl_NL'
         ? {
-            'multifunctionals.police_cars': <Toggle>{
-                type: 'toggle',
-                default: false,
-                dependsOn: '.vehicles.content',
-            },
-        }
+              'multifunctionals.police_cars': <Toggle>{
+                  type: 'toggle',
+                  default: false,
+                  dependsOn: '.vehicles.content',
+              },
+          }
         : null),
     ...(BUILD_LANG === 'de_DE'
         ? {
-            'multifunctionals.platform_trucks': <Toggle>{
-                type: 'toggle',
-                default: false,
-                dependsOn: '.vehicles.content',
-            },
-        }
+              'multifunctionals.platform_trucks': <Toggle>{
+                  type: 'toggle',
+                  default: false,
+                  dependsOn: '.vehicles.content',
+              },
+          }
         : null),
     'optionalAlternatives.allow_rw_instead_of_lf': <Toggle>{
         type: 'toggle',
@@ -112,16 +112,16 @@ export default ((MODULE_ID: string) => ({
     },
     ...(['de_DE', 'en_US', 'nl_NL', 'it_IT'].includes(BUILD_LANG)
         ? {
-            'optionalAlternatives.allow_ktw_instead_of_rtw': <Toggle>{
-                type: 'toggle',
-                default: false,
-                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                // @ts-ignore
-                disabled: (settings): boolean =>
-                    !settings[MODULE_ID]['patients.content'].value &&
-                    !settings[MODULE_ID]['vehicles.content'].value,
-            },
-        }
+              'optionalAlternatives.allow_ktw_instead_of_rtw': <Toggle>{
+                  type: 'toggle',
+                  default: false,
+                  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                  // @ts-ignore
+                  disabled: (settings): boolean =>
+                      !settings[MODULE_ID]['patients.content'].value &&
+                      !settings[MODULE_ID]['vehicles.content'].value,
+              },
+          }
         : null),
     'patients.title': <Toggle>{
         type: 'toggle',
@@ -137,20 +137,20 @@ export default ((MODULE_ID: string) => ({
     },
     ...(['en_GB', 'en_US', 'it_IT'].includes(BUILD_LANG)
         ? {
-            'patient.code_possible': <Toggle>{
-                type: 'toggle',
-                default: false,
-            },
-        }
+              'patient.code_possible': <Toggle>{
+                  type: 'toggle',
+                  default: false,
+              },
+          }
         : null),
     ...(BUILD_LANG !== 'nl_NL'
         ? {
-            'patients.patient_allow_first_responder_chance': <Toggle>{
-                type: 'toggle',
-                default: true,
-                dependsOn: '.patients.content',
-            },
-        }
+              'patients.patient_allow_first_responder_chance': <Toggle>{
+                  type: 'toggle',
+                  default: true,
+                  dependsOn: '.patients.content',
+              },
+          }
         : null),
     'patients.hideWhenNoNeed': <Toggle>{
         type: 'toggle',
@@ -187,11 +187,11 @@ export default ((MODULE_ID: string) => ({
     },
     ...(BUILD_LANG === 'en_US'
         ? {
-            'subsequent': <Toggle>{
-                type: 'toggle',
-                default: true,
-            },
-        }
+              subsequent: <Toggle>{
+                  type: 'toggle',
+                  default: true,
+              },
+          }
         : null),
     'followup': <Toggle>{
         type: 'toggle',
@@ -207,18 +207,18 @@ export default ((MODULE_ID: string) => ({
     },
     ...(BUILD_LANG === 'de_DE'
         ? {
-            k9_only_if_needed: <Toggle>{
-                type: 'toggle',
-                default: false,
-            },
-        }
+              k9_only_if_needed: <Toggle>{
+                  type: 'toggle',
+                  default: false,
+              },
+          }
         : null),
     ...(BUILD_LANG === 'nl_NL'
         ? {
-            bike_police_only_if_needed: <Toggle>{
-                type: 'toggle',
-                default: false,
-            },
-        }
+              bike_police_only_if_needed: <Toggle>{
+                  type: 'toggle',
+                  default: false,
+              },
+          }
         : null),
 })) as ModuleSettingFunction;
